Getting rid of the `join` field for `Cqrs.programWithFlags`
In trying to get rid of the `StateFact` intent the `Outcome Intent Fact`
approach really didn't work out well since the overall type of the program
would have to be `Program Flags Model (Outcome Intent Fact)` which has some
serious downsides:

* Views also inherit the type `Html (Outcome Intent Fact)` which would let you
state facts, the very reason for the experiment in the first place
* We wouldn't get around some kind of `Intent` to state facts since TEA needs
us to produce a `Cmd` of the type of message the program needs
* The muddied type signature would also be contagious to the outside world. One
goal of the overall experiment is that an outside consumer of the program
shouldn't know about the implementation details, which would leak through the
type signature.

What we did instead is to just remove the `join` requirement and let the user
explicitly produce a `Cmd Intent`, which means that in the end the user will
transparently use her own `StateFact` intent to retain the ability to just
produce facts from a `Consequence`. This solution is both more explicit and
less cumbersome.

Maybe we find a neat way to incorporate the `Outcome Intent Fact` without
polluting the overall program type.