Integrated `Vigors.Todo` but not without problems
The problem still remains that we need to find a sane way to pass
`Cmd msg`s from the init of a vigor to be dispatched by the embedding
program. Right now I just moved the `Cmd.none` into the `main` initialization.
It has to be possible to somehow bundle the boilerplate out of sight of the
client composing vigors.